tools/pyproject.toml
[project]
name = "cler-tools"
version = "0.1.0"
description = "Development tools for the Cler DSP flowgraph framework"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
cler-validate = "cler_tools.linter.validate:main"
cler-viz = "cler_tools.viz.visualize:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cler_tools"]